+++ src/lintFunctions.ts
@@ -61,6 +61,8 @@ export function runLint() {
 			if (stdout) {
 				logFunctions.writeLine(`${stdout}\n`, outputChannnel);
 				lintCurrentFile(stdout);
+				logFunctions.writeLine(`Delete file ${exeName}`, outputChannnel);
+				deleteFile(exeName, outputChannnel);
 			} else {
 				logFunctions.writeLine("No stdout from QB64.exe found", outputChannnel);
 			}
@@ -69,6 +71,23 @@ export function runLint() {
 	} catch (error) {
 		logFunctions.writeLine(`ERROR in runLint: ${error}`, outputChannnel);
 	}
+}
+
+/**
+ * Deletes a file see (https://stackoverflow.com/questions/5315138/node-js-remove-file)
+ * @param fileName {string} File to delete
+ * @returns void
+ */
+function deleteFile(fileName: string, outputChannnel: any) {
+	const { unlink } = require('fs/promises');
+	(async function (path) {
+		try {
+			await unlink(path);
+			logFunctions.writeLine(`File ${path} Deleted`, outputChannnel)
+		} catch (error) {
+			logFunctions.writeLine(`ERROR in deleteFile: ${error.message}`, outputChannnel)
+		}
+	})(fileName);
 }
 
 /**
@@ -102,29 +121,29 @@ function lintCurrentFile(compilerOutput: string) {
 			}
 
 			errorLineNumber = -1;
-			if (lintLine.startsWith("Illegal")
-				|| lintLine.startsWith("DIM: Expected")
-				|| lintLine.startsWith("Expected =")
-				|| lintLine.startsWith("Cannot convert type")
-				|| lintLine.startsWith("Undefined")
+			if (lintLine.startsWith("Illegal ")
+				|| lintLine.startsWith("DIM: ")
+				|| lintLine.startsWith("Cannot ")
+				|| lintLine.startsWith("Undefine d")
 				|| lintLine.startsWith("Expected")
-				|| lintLine.startsWith("File '")
-				|| lintLine.startsWith("Syntax error")
-				|| lintLine.startsWith("RETURN linelabel")
-				|| lintLine.startsWith("Type symbols after")
-				|| lintLine.startsWith("Name already in use")
-				|| lintLine.startsWith("Variable")
-				|| lintLine.startsWith("Unexpected")
+				|| lintLine.startsWith("File ")
+				|| lintLine.startsWith("Syntax ")
+				|| lintLine.startsWith("RETURN ")
+				|| lintLine.startsWith("Type ")
+				|| lintLine.startsWith("Name ")
+				|| lintLine.startsWith("Unexpected ")
 				|| lintLine.startsWith("Invalid expression")
 				|| lintLine.startsWith("Element not defined")
-				|| lintLine.startsWith("Name already in use")
-				|| lintLine.startsWith("Unknown type")
-				|| lintLine.startsWith("Missing (")
+				|| lintLine.startsWith("Unknown ")
+				|| lintLine.startsWith("Missing ")
 				|| lintLine.startsWith("_DEFINE: ")
-				|| lintLine.startsWith("Command not implemented")
+				|| lintLine.startsWith("Command ")
 				|| lintLine.startsWith("2nd sub argument")
-				|| lintLine.startsWith("Cannot convert")
+				|| lintLine.startsWith("Cannot ")
 				|| lintLine.startsWith("Invalid ")
+				|| lintLine.startsWith("Variable")
+				|| lintLine.startsWith("Array")
+				|| lintLine.startsWith("THEN ")
 			) {
 				let code: string = "";
 				for (let x = lineIndex; x < lines.length; x++) {
@@ -163,6 +182,7 @@ function lintCurrentFile(compilerOutput: string) {
 				const tokens: string[] = lintLine.split(":");
 
 				if (path.basename(document.uri.fsPath) != tokens[0]) {
+					// Somehow highlight the file in the explorer view and maybel the include statement that goes with it.
 					continue;
 				}
 				errorLineNumber = Number(tokens[1]) - 1;